+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 <!-- Hi. -->
 
-## tdl@7.0.0 (unreleased)
+## tdl@7.0.0 (2021-01-14)
 
 - **Important**: TypeScript and Flow users now need to install `tdlib-types` for the TDLib typings to work.
 It is now possible to install the typings for other TDLib versions.
@@ -15,14 +15,14 @@ It is now possible to install the typings for other TDLib versions.
 - Internal:
 - - Updated Flow to v0.138.0.
 
-## tdlib-types@1 (unreleased)
+## tdlib-types@1 (2021-01-14)
 
 - This is the first release of `tdlib-types`, see its [README](packages/tdlib-types/README.md).
 - The TDLib typings have been split from the `tdl` package into this package.
 - The generator now parses "may be null" comments.
 - Added typings for TDLib v1.7.0.
 
-## tdl-shared@0.10.0, tdl-tdlib-addon@1.0.0, tdl-tdlib-ffi@3.0.0, tdl-tdlib-wasm@0.6.0 (unreleased)
+## tdl-shared@0.10.0, tdl-tdlib-addon@1.0.0, tdl-tdlib-ffi@3.0.0, tdl-tdlib-wasm@0.6.0 (2021-01-14)
 
 - Removed deprecated `setLogFilePath`, `setLogMaxFileSize`, and `setLogVerbosityLevel`.
 - New method: `getName(): string`.